Fix bug in blackrock RNG to avoid it RNG'ing in groups
This fixes a issue where blackrock would work, but would only generate
groups of numbers at a time. This is quite annoying if you are trying
to use blackrock to do scanning, because it means that some areas
are going to be hit MUCH harder than others, then it will move on
to the next area and hit that harders.

It was still better than doing it sequentially, but it is not
what blackrock was designed to do.

This was bug was caused by type confusion while porting it from
C code, and was picked up because the RDNS scanner I was writing
was not scanning some IP blocks.